Changed connection pool ,Switch removed
The switch picked the Properties file to load, but this is now done in
the Model where a similar switch determines the DAO types depending on
the DB type (MySql SQL_Server or Test.

The connectionPool is now a LAzy singleton, so the Properties can be
set by the Model before asking for a connectionPool instance.